+++ src/api/tag.rs
@@ -5,11 +5,11 @@ pub struct Tag {
     pub name: String,
     pub digest: String,
     pub created: DateTime<Utc>,
-    pub size: u32
+    pub size: u64
 }
 
 impl Tag {
-    pub fn new(name: String, digest: String, created: DateTime<Utc>, size: u32) -> Self {
+    pub fn new(name: String, digest: String, created: DateTime<Utc>, size: u64) -> Self {
         Self { name, digest, created, size }
     }
 }
